# lib/api-zod/src/generated/types/cartItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tanim Agricultural Supply Chain API
 * OpenAPI spec version: 0.1.0
 */

export interface CartItem {
  id: number;
  product_id: number;
  product_name: string;
  product_image?: string | null;
  price: number;
  quantity: number;
  stock_quantity: number;
  unit: string;
  farmer_name: string;
  subtotal: number;
}