# lib/api-zod/src/generated/types/updateCartItemRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tanim Agricultural Supply Chain API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateCartItemRequest {
  /** @minimum 1 */
  quantity: number;
}